+++ src/db/mod.rs
@@ -1,5 +1,6 @@
 use std::sync::atomic::{AtomicU32, Ordering};
 
+use crate::accumulate::Accumulate;
 use crate::cell::CellData;
 use crate::storage::{ComputationId, StorageFor};
 use crate::{Cell, OutputType, Storage};
@@ -38,16 +39,12 @@ impl<S: Default> Default for Db<S> {
 
 /// Abstracts over the `get` function provided by `Db<S>` and `DbHandle<S>` to avoid
 /// providing `get` and `get_db` variants for each function.
-#[cfg(not(feature = "async"))]
 pub trait DbGet<C: OutputType> {
+    /// Run an incremental computation `C` and return its output.
+    /// If `C` is already cached, no computation will be performed.
     fn get(&self, key: C) -> C::Output;
 }
-#[cfg(feature = "async")]
-pub trait DbGet<C: OutputType> {
-    fn get(&self, key: C) -> impl Future<Output = C::Output> + Send;
-}
 
-#[cfg(not(feature = "async"))]
 impl<S, C> DbGet<C> for Db<S>
 where
     C: OutputType + ComputationId,
@@ -55,17 +52,6 @@ where
 {
     fn get(&self, key: C) -> C::Output {
         self.get(key)
-    }
-}
-
-#[cfg(feature = "async")]
-impl<S, C> DbGet<C> for Db<S>
-where
-    C: OutputType + ComputationId,
-    S: Storage + StorageFor<C> + Sync,
-{
-    fn get(&self, key: C) -> impl Future<Output = C::Output> + Send {
-        Db::get(self, key)
     }
 }
 
@@ -165,7 +151,6 @@ impl<S: Storage> Db<S> {
     } 
 }
 
-#[cfg(not(feature = "async"))]
 impl<S: Storage> Db<S> {
     /// Updates an input with a new value
     ///
@@ -333,163 +318,53 @@ impl<S: Storage> Db<S> {
     fn with_cell<R>(&self, cell: Cell, f: impl FnOnce(&CellData) -> R) -> R {
         f(&self.cells.get(&cell).unwrap())
     }
-}
 
-#[cfg(feature = "async")]
-impl<S: Storage + Sync> Db<S> {
-    /// Updates an input with a new value
+    /// Fill the given set with all cells directly or indirectly used as dependencies
+    /// for the given cell. To ensure dependencies are up to date, this may run
+    /// any stale dependencies.
+    fn collect_all_dependencies(&self, operation: Cell) -> Vec<Cell> {
+        // Ensure all dependencies are up to date
+        self.update_cell(operation);
+
+        let mut queue: Vec<_> = self.with_cell(operation, |cell| {
+            cell.dependencies.iter().copied().collect()
+        });
+
+        let mut cells = Vec::new();
+        cells.push(operation);
+
+        // We must be careful to preserve dependency order in `cells`.
+        // This is easiest if we push from last-to-first and reverse `cells` later.
+        while let Some(dependency) = queue.pop() {
+            cells.push(dependency);
+
+            self.with_cell(dependency, |cell| {
+                for dependency in cell.dependencies.iter() {
+                    queue.push(*dependency);
+                }
+            });
+        }
+
+        cells.reverse();
+        cells
+    }
+
+    /// Retrieve an accumulated value in a container of the user's choice.
+    /// This will return all the accumulated items after the given computation.
     ///
-    /// Panics in debug mode if the input is not an input - ie. it has at least 1 dependency.
-    /// Note that this check is skipped when compiling in Release mode.
-    pub async fn update_input<C: OutputType>(&mut self, input: C, new_value: C::Output)
-    where
-        C: ComputationId,
-        S: StorageFor<C>,
-    {
-        let cell_id = self.get_or_insert_cell(input);
-        debug_assert!(
-            self.is_input(cell_id).await,
-            "`update_input` given a non-input value. Inputs must have 0 dependencies",
-        );
-
-        let changed = self.storage.update_output(cell_id, new_value);
-        let mut cell = self.cells.get(&cell_id).unwrap();
-
-        if changed {
-            let version = self.version.fetch_add(1, Ordering::SeqCst) + 1;
-            cell.last_updated_version = version;
-            cell.last_verified_version = version;
-        } else {
-            cell.last_verified_version = self.version.load(Ordering::SeqCst);
-        }
-    }
-
-    fn is_input(&self, cell: Cell) -> impl Future<Output = bool> {
-        self.with_cell(cell, |cell| cell.dependencies.len() == 0)
-    }
-
-    /// True if a given input is stale and needs to be re-computed.
-    /// Inputs which have never been computed are also considered stale.
+    /// Note that although this method will not re-perform the given computation,
+    /// it will re-collect all the required accumulated items each time it is called,
+    /// which may be costly for large dependency trees.
     ///
-    /// This does not actually re-compute the input.
-    pub async fn is_stale<C: OutputType>(&self, input: &C) -> bool
-    where
-        S: StorageFor<C>,
-    {
-        // If the cell doesn't exist, it is definitely stale
-        let Some(cell) = self.get_cell(input) else {
-            return true;
-        };
-        self.is_stale_cell(cell).await
-    }
-
-    /// True if a given cell is stale and needs to be re-computed.
-    ///
-    /// Note that this may re-compute some input
-    async fn is_stale_cell(&self, cell: Cell) -> bool {
-        let computation_id = self.with_cell(cell, |data| data.computation_id).await;
-
-        if self.storage.output_is_unset(cell, computation_id) {
-            return true;
-        }
-
-        // if any dependency may have changed, this cell is stale
-        let (last_verified, dependencies) = self
-            .with_cell(cell, |data| {
-                (data.last_verified_version, data.dependencies.clone())
-            })
-            .await;
-
-        // Dependencies need to be iterated in the order they were computed.
-        // Otherwise we may re-run a computation which does not need to be re-run.
-        // In the worst case this could even lead to panics - see the div0 test.
-        for dependency_id in dependencies {
-            self.update_cell(dependency_id).await;
-
-            // This cell is stale if the dependency has been updated since
-            // we last verified this cell
-            if self
-                .cells
-                .read(&dependency_id, |_, dependency| {
-                    dependency.last_updated_version > last_verified
-                })
-                .unwrap()
-            {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Similar to `update_input` but runs the compute function
-    /// instead of accepting a given value. This also will not update
-    /// `self.version`
-    async fn run_compute_function(&self, cell_id: Cell) {
-        let computation_id = self.with_cell(cell_id, |data| data.computation_id).await;
-
-        let handle = self.handle(cell_id);
-        let changed = S::run_computation(&handle, cell_id, computation_id).await;
-
-        let version = self.version.load(Ordering::SeqCst);
-        let mut cell = self.cells.get_async(&cell_id).await.unwrap();
-        cell.last_verified_version = version;
-
-        if changed {
-            cell.last_updated_version = version;
-        }
-    }
-
-    /// Trigger an update of the given cell, recursively checking and re-running any out of date
-    /// dependencies.
-    async fn update_cell(&self, cell_id: Cell) {
-        let last_verified_version = self
-            .with_cell(cell_id, |data| data.last_verified_version)
-            .await;
-        let version = self.version.load(Ordering::SeqCst);
-
-        if last_verified_version != version {
-            // if any dependency may have changed, update
-            if Box::pin(self.is_stale_cell(cell_id)).await {
-                self.run_compute_function(cell_id).await;
-            } else {
-                let mut cell = self.cells.get_async(&cell_id).await.unwrap();
-                cell.last_verified_version = version;
-            }
-        }
-    }
-
-    /// Retrieves the up to date value for the given computation, re-running any dependencies as
-    /// necessary.
-    ///
-    /// This function can panic if the dynamic type of the value returned by `compute.run(..)` is not `T`.
-    pub fn get<C: OutputType + ComputationId>(&self, compute: C) -> impl Future<Output = C::Output>
-    where
-        S: StorageFor<C>,
+    /// This is most often used for operations like retrieving diagnostics or logs.
+    pub fn get_accumulated<Container, Item, C>(&self, compute: C) -> Container where
+        Container: FromIterator<Item>,
+        S: Accumulate<Item> + StorageFor<C>,
+        C: OutputType + ComputationId
     {
         let cell_id = self.get_or_insert_cell(compute);
-        self.get_with_cell::<C>(cell_id)
-    }
 
-    pub(crate) fn get_with_cell<Concrete: OutputType>(
-        &self,
-        cell_id: Cell,
-    ) -> impl Future<Output = Concrete::Output> + Send
-    where
-        S: StorageFor<Concrete>,
-    {
-        async move {
-            self.update_cell(cell_id).await;
-
-            self.storage
-                .get_output(cell_id)
-                .expect("cell result should have been computed already")
-        }
-    }
-
-    async fn with_cell<R>(&self, cell: Cell, f: impl FnOnce(&CellData) -> R) -> R {
-        self.cells
-            .read_async(&cell, |_, data| f(data))
-            .await
-            .unwrap()
+        let cells = self.collect_all_dependencies(cell_id);
+        self.storage.get_accumulated(&cells)
     }
 }