+++ tests/compiler.rs
@@ -1,4 +1,3 @@
-#![cfg(not(feature = "async"))]
 use std::{
     collections::{BTreeMap, BTreeSet},
     rc::Rc,